% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dpt.R
\name{read_dpt}
\alias{read_dpt}
\title{load a .dpt file}
\usage{
read_dpt(path_to_file)
}
\arguments{
\item{path_to_file}{path to the input file}
}
\value{
A list of the input file
}
\description{
This function reads in a .dpt file and returns a list.
}
\examples{
read_dpt(path_to_file = "workspace/data.dpt")
}
