% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_to_csv.R
\name{spectra_to_csv}
\alias{spectra_to_csv}
\title{Write spectral data to CSV}
\usage{
spectra_to_csv(spectra, out_path, out_file)
}
\arguments{
\item{spectra}{list object to transform}

\item{out_path}{path to directory if outputs}

\item{out_file}{filename basestring of outputs}
}
\value{
2 CSV files
}
\description{
Takes a .dpt list object and calculates
the number of particles and columns in the data.
Renames the columns, transforms to long format and
writes 2 CSV, 1 in long and 1 in wide format.
}
\examples{
spectra_to_csv(spectra = spectra, out_path = "C:/workspace/", out_file = "test_out")
}
