% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_to_long.R
\name{spectra_to_long}
\alias{spectra_to_long}
\title{Transforms spectral data to long format}
\usage{
spectra_to_long(spectra_wide)
}
\arguments{
\item{spectra_wide}{list object to transform}
}
\value{
spectra list object in wide format
}
\description{
Takes a .dpt list input and transforms to long.
}
\examples{
spectra_to_long(spectra_wide = spectra)
}
